requirements.txt
django-cors-headers==2.5.2
Django==2.2.3
djangorestframework==3.9.4
psycopg2==2.8.2
requests==2.22.0